% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeReport.R
\name{arrangeCompletenessOutput}
\alias{arrangeCompletenessOutput}
\title{This function arrange the completeness output of the interested 
genome into the corresponding folder in the output folder}
\usage{
arrangeCompletenessOutput(
  completenessOutput,
  output,
  coreSet,
  genomeName,
  scoreMode
)
}
\arguments{
\item{output}{The directory which contains the output directory}

\item{coreSet}{The name of the interested core set. The core directory can
contains more than one core set and the user must specify the interested
core set. The core set will be stored in the folder core_orthologs in
subfolder, specify them by the name of the subfolder}

\item{genomeName}{the name of the interested genome}

\item{scoreMode}{the mode determines the method to scoring the founded
ortholog and how to classify them. Choices: 1, 2, 3, "len"}

\item{completenessOuput}{a list contains the completeness output. The full
table, missing table and the ignored table}
}
\description{
This function arrange the completeness output of the interested 
genome into the corresponding folder in the output folder
}
