% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFdog.R
\name{arrangePPOutput}
\alias{arrangePPOutput}
\title{This function arrange the phyloprofile output of the interested 
genome into the corresponding folder in the output folder}
\usage{
arrangePPOutput(
  ppOutputList,
  refSpecList,
  output,
  coreSet,
  genomeName,
  scoreMode,
  extend
)
}
\arguments{
\item{ppOutputList}{a list contains the phylogenetic profile of the
interested genome in data.frame, the fasta file of the genome in vector,
2 file domains of the genome in data.frame (two file domains can be NULL)
#' @param refSpecList A list contains one or many genome ID of the genomes,
which were used to build the core set. The genome ID of this list will be
stored with an priority order, the tool look at into the fasta file of each
core group and determine with the priority order to determine the references
species for each core group.}

\item{output}{The directory which contains the output directory}

\item{coreSet}{The name of the interested core set. The core directory can
contains more than one core set and the user must specify the interested
core set. The core set will be stored in the folder core_orthologs in
subfolder, specify them by the name of the subfolder}

\item{genomeName}{the name of the interested genome}

\item{scoreMode}{the mode determines the method to scoring the founded
ortholog and how to classify them. Choices: 1, 2, 3, "busco"}

\item{extend}{The output of the function is a phylogenetic profile of the
interested genome. It contains 4 files, .phyloprofile, .extended.fa,
_reverse.domains and _forward.domains. If extend = TRUE, the files will be
appended into the old files in the folder output of the core directory or in
the inputed folder by the user with the argument ppDir. If there is no old
files in the folder, the output files of the function will be writen in the
new files.}
}
\description{
This function arrange the phyloprofile output of the interested 
genome into the corresponding folder in the output folder
}
