% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlRefSpec.R
\name{autofindPriority}
\alias{autofindPriority}
\title{The function determines the priority order of a list of taxa based on the
taxonomic distance of the taxa in the list with a given taxon}
\usage{
autofindPriority(query, refSet)
}
\arguments{
\item{query}{The genome ID of the interested genome}

\item{refSet}{the list of taxa in the core set}
}
\value{
The list of taxa with a priority order based on the taxanomic
distance
}
\description{
The function determines the priority order of a list of taxa based on the
taxonomic distance of the taxa in the list with a given taxon
}
