% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCoreSet.R
\name{calculateAllCutoff}
\alias{calculateAllCutoff}
\title{The function calculate cut off values for all core genes in the set}
\usage{
calculateAllCutoff(root, coreSet)
}
\arguments{
\item{root}{the path to the root folder}

\item{coreSet}{the core set name}
}
\value{
none
}
\description{
The function calculate cut off values for all core genes in the set
}
