% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{calculateBuscoCutoff}
\alias{calculateBuscoCutoff}
\title{The function to calculate the cutoff value for the busco mode}
\usage{
calculateBuscoCutoff(coreSet, coreGene)
}
\arguments{
\item{coreSet}{the path to the core set}

\item{coreGene}{the ID of the core gene}
}
\value{
a list that contains the mean length and the standard deviation of
the length of the core gene
}
\description{
The function to calculate the cutoff value for the busco mode
}
