% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCompleteness.R
\name{checkArguments}
\alias{checkArguments}
\title{Check if all the input are correct}
\usage{
checkArguments(
  genome,
  fasAnno = NULL,
  root,
  coreSet,
  extend = FALSE,
  redo = FALSE,
  scoreMode,
  priorityList = NULL,
  cpu = 4
)
}
\arguments{
\item{genome}{The path to the genome fasta file}

\item{fasAnno}{The path to the fas annotation file. Can be NULL}

\item{root}{the path to the root folder}

\item{coreSet}{The core set name}

\item{extend}{A logical value}

\item{redo}{A logical value}

\item{scoreMode}{The mode determines the method to assess the orthologs}

\item{priorityList}{The list of taxa in the core set to determine the
references species}

\item{cpu}{The number of the cores that HaMStR will use}
}
\value{
A list that contains a logical value and the message to the error
}
\description{
Check if all the input are correct
}
