% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCompleteness.R
\name{checkCompleteness}
\alias{checkCompleteness}
\title{The function to check completeness of a new given genome}
\usage{
checkCompleteness(
  genome,
  fasAnno = NULL,
  coreSet,
  extend = FALSE,
  redo = FALSE,
  scoreMode,
  priorityList = NULL,
  cpu = 4
)
}
\arguments{
\item{genome}{The path to the fasta file of the genome}

\item{fasAnno}{the path to the json file of the annotation of the genome. If
equal NULL the function will create one}

\item{coreSet}{the path to the core set}

\item{extend}{the logical option to decide if the information of the
interested will be appended to the original files and reused for the next
check}

\item{redo}{the logical option to decide if the tool should recheck for a
genome ID that already exists in the original pp}

\item{scoreMode}{the mode determines the method to assess the founded
ortholog}

\item{priorityList}{the priority list to determine the references species}

\item{cpu}{the number of the cores}
}
\value{
a list that contains a report table of the completeness of the
interested genome and a table that contains the report of the interested
genome within the other taxa in the original pp
}
\description{
The function to check completeness of a new given genome
}
