% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCompleteness.R
\name{checkExist}
\alias{checkExist}
\title{Check if the interested genome exists already in the original pp}
\usage{
checkExist(genomeName, coreSet, scoreMode)
}
\arguments{
\item{genomeName}{the ID of the genome}

\item{coreSet}{the path to the core set}

\item{scoreMode}{the mode determines the method to assess the ortholog}
}
\value{
TRUE or FALSE
}
\description{
Check if the interested genome exists already in the original pp
}
