% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeOriginal.R
\name{computeOriginal}
\alias{computeOriginal}
\title{The function compute the initial original phylogenetic profile of all the
taxa in the score set}
\usage{
computeOriginal(root, coreSet, scoreMode, cpu = 4, cleanup)
}
\arguments{
\item{root}{the path to the root folder}

\item{coreSet}{the core set name}

\item{scoreMode}{the mode determines the way to assess the founded ortholog}

\item{cpu}{determines the number of the cores}
}
\value{
none
}
\description{
The function compute the initial original phylogenetic profile of all the
taxa in the score set
}
