% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeReport.R
\name{computeReport}
\alias{computeReport}
\title{The function create the report of the completeness of an interested genome}
\usage{
computeReport(
  genome,
  fasAnno,
  root,
  coreSet,
  extend = FALSE,
  scoreMode,
  priorityList = NULL,
  cpu,
  computeOri = FALSE
)
}
\arguments{
\item{genome}{the path to the fasta file of the interested genome}

\item{fasAnno}{the path to the json file of the fas annotation of the genome}

\item{coreSet}{the path to the core set}

\item{extend}{the logical argument to decision if the information of the
interested genome will be appended to the original files}

\item{scoreMode}{the mode determines the method to assess the founded
ortholog}

\item{priorityList}{the priority list to determines the references species
If equal NULL the tool will create the list based on the taxonomic distance}

\item{cpu}{determines the number of the cores}

\item{computeOri}{a logical argument to determine the special case when the
function was used as a sub function for the function computeOriPP}
}
\value{
none
}
\description{
The function create the report of the completeness of an interested genome
}
