% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFdog.R
\name{concanateFiles}
\alias{concanateFiles}
\title{After fdogFAS is finished. It will leave in a temporary folder the 
phylogenetic profiles (pp). The pps can not be used direct to assess the
completeness of the interested genome. This function processes the pp files
in a folder and merged them.}
\usage{
concanateFiles(directory, genomeName, scoreMode)
}
\arguments{
\item{directory}{path to the folder, that contains the pp files}

\item{genomeName}{the genomeID of the interested genome}

\item{scoreMode}{the mode determines the method to scoring the founded 
ortholog and how to classify them. Choices: 1, 2, 3, "busco"}
}
\value{
a list which contains the pp of the interested genome in data.frame, 
the extended fasta file of the interested genome in form of a vector, which
contains the lines of the file and 2 domains file in form of data.frame
}
\description{
After fdogFAS is finished. It will leave in a temporary folder the 
phylogenetic profiles (pp). The pps can not be used direct to assess the
completeness of the interested genome. This function processes the pp files
in a folder and merged them.
}
\examples{
## Creating some pseudo data
geneID <- c("HUMAN@96063")
ncbiID <- c("ncbi9606")
orthoID <- c("123at123|HUMAN@9606@3|Q123123|1")
FAS_F <- c(1)
FAS_B <- c(1)
pp1 <- data.frame(geneID, ncbiID, orthoID, FAS_F, FAS_B)
pp2 <- data.frame(geneID, ncbiID, orthoID, FAS_F, FAS_B)

## Write them in 2 files in a test folder in the current working directory
wd <- getwd()
testFolder <- paste(wd, "/fCAT_testfunction", sep = "")
dir.create(testFolder, recursive = TRUE)
filePath1 <- paste(testFolder, "/test1.phyloprofile", sep = "")
filePath2 <- paste(testFolder, "/test2.phyloprofile", sep = "")
write.table(pp1, filePath1, row.names = FALSE, quote = FALSE, sep = "\t")
write.table(pp2, filePath2, row.names = FALSE, quote = FALSE, sep = "\t")

test <- concanateFiles(testFolder, "HUMAN@9606@3", scoreMode = 2)
print(test)

## Delete test folder
unlink(testFolder, recursive = TRUE)
}
