% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFiles.R
\name{correctFiles}
\alias{correctFiles}
\title{remove all the lines of a specific genome from all of the phyloprofile,
domains, fasta files, priority file, and report file in a folder}
\usage{
correctFiles(directory, genome)
}
\arguments{
\item{directory}{the path to the directory, that contains the files}

\item{genome}{name of the genome, that should be removed}
}
\value{
none
}
\description{
remove all the lines of a specific genome from all of the phyloprofile,
domains, fasta files, priority file, and report file in a folder
}
