% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFiles.R
\name{correctPP}
\alias{correctPP}
\title{remove the lines of a specific genome from a phylogenetic profile}
\usage{
correctPP(PPFile, genome)
}
\arguments{
\item{PPFile}{path to phylogenetic profile}

\item{genome}{name of the genome, that should be removed}
}
\value{
none
}
\description{
remove the lines of a specific genome from a phylogenetic profile
}
