% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFiles.R
\name{correctPriority}
\alias{correctPriority}
\title{If user choose the option to extend the phylogenetic profile of the
interested genome into the original phylogenetic profile, the tool must save
the priority list that the user entered to use this one to get the result for
the next calling. When the user want to redo the check, the function will
remove the old taxa in teh priorityFile}
\usage{
correctPriority(priorityFile, genome)
}
\arguments{
\item{priorityFile}{The path to the text file, that contains the information
about the priority list of all searching}

\item{genome}{the genome ID that need to be remove}
}
\value{
none
}
\description{
If user choose the option to extend the phylogenetic profile of the
interested genome into the original phylogenetic profile, the tool must save
the priority list that the user entered to use this one to get the result for
the next calling. When the user want to redo the check, the function will
remove the old taxa in teh priorityFile
}
