% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFiles.R
\name{correctReport}
\alias{correctReport}
\title{Remove a specific genome from the report file}
\usage{
correctReport(reportFile, genome)
}
\arguments{
\item{reportFile}{the path to the report file}

\item{genome}{the genome ID that need to be removed}
}
\value{
none
}
\description{
Remove a specific genome from the report file
}
