% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCoreSet.R
\name{createAllAnnotation}
\alias{createAllAnnotation}
\title{The function creates all FAS annotation for all core genes in the set}
\usage{
createAllAnnotation(coreSet)
}
\arguments{
\item{coreSet}{the path to the core set}
}
\value{
none
}
\description{
The function creates all FAS annotation for all core genes in the set
}
