% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCoreSet.R
\name{createAnnotation}
\alias{createAnnotation}
\title{The function creates FAS annotation for a core gene in the core set}
\usage{
createAnnotation(coreSet, coreGene)
}
\arguments{
\item{coreSet}{The path to the core set}

\item{coreGene}{The name of the core gene in the set}
}
\value{
none
}
\description{
The function creates FAS annotation for a core gene in the core set
}
