% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFasScore.R
\name{extendDomain}
\alias{extendDomain}
\title{The function correct information in a original domain file. This function is
just applied when scoreMode=1}
\usage{
extendDomain(domain)
}
\arguments{
\item{domain}{The path to the domain file}
}
\value{
none
}
\description{
The function correct information in a original domain file. This function is
just applied when scoreMode=1
}
