% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHamstr.R
\name{extendOriginal}
\alias{extendOriginal}
\title{Function to append the phylogenetic profile of the interested genome into the
original pp}
\usage{
extendOriginal(root, coreSet, scoreMode, fileList)
}
\arguments{
\item{root}{the path to the root folder}

\item{coreSet}{the core set name}

\item{scoreMode}{the mode to assess the founded ortholog}

\item{fileList}{a list that contains the information of the phylogenetic
profile of the interested genome}
}
\description{
Function to append the phylogenetic profile of the interested genome into the
original pp
}
