% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenome.R
\name{extractDomains}
\alias{extractDomains}
\title{A domains file can contains lines of many different genomes. This function
extract the domains of a specific genome}
\usage{
extractDomains(domains, genome, reverse = FALSE)
}
\arguments{
\item{domains}{The domains in data.frame}

\item{genome}{The genome ID of the genome, whose domains need to be extracted}

\item{reverse}{if reverse = TRUE then the function will not extract the lines
of the genome but remove it}
}
\value{
the domains of the interested genome in data.frame
}
\description{
A domains file can contains lines of many different genomes. This function
extract the domains of a specific genome
}
