% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenome.R
\name{extractPP}
\alias{extractPP}
\title{This function takes a phylogenetic profile, which contains many different
genomes as an input and extract the phylogenetic profile of a specific genome}
\usage{
extractPP(pp, genome)
}
\arguments{
\item{pp}{A phylogenetic profile in data.frame}

\item{genome}{The genome ID of the genome, whose pp need to be extracted.
Exp: genome = HUMAN@9606@3}
}
\value{
The phylogenetic profile of the interested genome in data.frame
}
\description{
This function takes a phylogenetic profile, which contains many different
genomes as an input and extract the phylogenetic profile of a specific genome
}
\examples{
## Create pseudo phylogenetic profile, which contains two different genomes
geneID <- c("530670", "530730", "603043")
ncbiID <- c("ncbi9606", "ncbi9606", "ncbi9606")
orthoID <- c("530670|HUMAN@9606@3|Q16526|1", "530730|HUMAN@9606@3|P05091|1",
"603043|HOMSA@9606@2|Q39233|1")
FAS_F <- c(1, 1, 1)
FAS_B <- c(1, 1, 1)
pp <- data.frame(geneID, ncbiID, orthoID, FAS_F, FAS_B)
newPP <- extractPP(pp, "HUMAN@9606@3")
print.data.frame(newPP)
}
