% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFasScore.R
\name{getFasScore}
\alias{getFasScore}
\title{The function calculate the FAS score in the scoreMode 1 and update the scores
into phylogenetic profile of a specific core gene}
\usage{
getFasScore(coreSet, coreGene, extend = FALSE, refSpec)
}
\arguments{
\item{coreSet}{the path to the core set}

\item{coreGene}{the ID of the core gene}

\item{extend}{the logical argument to decide if the information of the
interested genome will be appended to the original files}

\item{refSpec}{the references species}
}
\value{
none
}
\description{
The function calculate the FAS score in the scoreMode 1 and update the scores
into phylogenetic profile of a specific core gene
}
