% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlRefSpec.R
\name{getSpec}
\alias{getSpec}
\title{The function read the fasta file in each core gene and extract the references
species based on the priority order in the priorityList}
\usage{
getSpec(fasta, priorityList)
}
\arguments{
\item{fasta}{the path to the fasta file}

\item{priorityList}{a vector that contains the taxa of the core set (it does
not need to contain all the taxa, can be all or fewer) in a priority list}
}
\value{
The function will return the references species. If the function did
not find any references species, it will return the value NULL
}
\description{
The function read the fasta file in each core gene and extract the references
species based on the priority order in the priorityList
}
