% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCoreSet.R
\name{processCoreSet}
\alias{processCoreSet}
\title{The function compute the FAS annotations and calculate the cut off values for
all core genes in the set}
\usage{
processCoreSet(root, coreSet)
}
\arguments{
\item{root}{the path to the root folder}

\item{coreSet}{the core set name}
}
\value{
none
}
\description{
The function compute the FAS annotations and calculate the cut off values for
all core genes in the set
}
