% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{reportSingle}
\alias{reportSingle}
\title{Create the report of the completeness of a genome based on its phylogenetic
profile}
\usage{
reportSingle(pp, coreSet, scoreMode, priorityList)
}
\arguments{
\item{pp}{the phylogenetic profile of the genome in data frame}

\item{coreSet}{the path to the core set}

\item{scoreMode}{the mode to determined the method to assess the ortholog}

\item{priorityList}{the list to determinde the references species}
}
\value{
the report in data frame
}
\description{
Create the report of the completeness of a genome based on its phylogenetic
profile
}
