% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFasScore.R
\name{updateFasScore}
\alias{updateFasScore}
\title{The function calculate the FAS score in the scoreMode 1 and update the scores
into phylogenetic profile of a specific core gene}
\usage{
updateFasScore(root, coreSet, coreGene, extend, refSpec)
}
\arguments{
\item{root}{Path to the root folder}

\item{coreSet}{The core set name}

\item{coreGene}{the ID of the core gene}

\item{extend}{the logical argument to decide if the information of the
interested genome will be appended to the original files}

\item{refSpec}{the references species}
}
\value{
none
}
\description{
The function calculate the FAS score in the scoreMode 1 and update the scores
into phylogenetic profile of a specific core gene
}
