% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busco.R
\name{updateLength}
\alias{updateLength}
\title{In the scoreMode "busco" the tool must calculate the length of the founded
ortholog to perfome the assessment. After HaMStR founded an ortholog the
function will be called to update the length of the ortholog in phylogenetic
profile}
\usage{
updateLength(coreSet, coreGene)
}
\arguments{
\item{coreSet}{the path to the core set}

\item{coreGene}{the ID of the core gene}
}
\value{
none
}
\description{
In the scoreMode "busco" the tool must calculate the length of the founded
ortholog to perfome the assessment. After HaMStR founded an ortholog the
function will be called to update the length of the ortholog in phylogenetic
profile
}
