% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{assessBusco}
\alias{assessBusco}
\title{The function to assess the founded ortholog with the algorithm of busco}
\usage{
assessBusco(orthoLength, root, coreSet, coreGene, f)
}
\arguments{
\item{orthoLength}{the length of the ortholg sequence}

\item{root}{the path to the root folder}

\item{coreSet}{the core set name}

\item{coreGene}{the ID of the core gene}

\item{f}{the frequent of the core gene in the pp}
}
\value{
the status of the core gene
}
\description{
The function to assess the founded ortholog with the algorithm of busco
}
