% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{assessStatus}
\alias{assessStatus}
\title{The function to assess the status of a founded ortholog}
\usage{
assessStatus(fasF, fasB, root, coreSet, coreGene, scoreMode, f, priorityList)
}
\arguments{
\item{fasF}{the forward fas score of the ortholog}

\item{fasB}{the backward fas score of the ortholog}

\item{root}{the path to the root folder}

\item{coreSet}{the core set name}

\item{coreGene}{the ID of the core gene}

\item{scoreMode}{the mode to determines the method to assess the ortholog}

\item{f}{the frequent of the core genes in the pp}

\item{priorityList}{the priority list to determines the references species}
}
\value{
the status of the ortholog of the core gene
}
\description{
The function to assess the status of a founded ortholog
}
