% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlRefSpec.R
\name{autofindPriority}
\alias{autofindPriority}
\title{The function determines the priority order of a list of taxa based on the
taxonomic distance between the taxa in the list and a given taxon}
\usage{
autofindPriority(query, refSet)
}
\arguments{
\item{query}{The genome ID of the interested genome}

\item{refSet}{the list of genome ID of the core genomes}
}
\value{
The list of core genomes'ID with a priority order based on the
taxanomic distance
}
\description{
The function determines the priority order of a list of taxa based on the
taxonomic distance between the taxa in the list and a given taxon
}
