% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCoreSet.R
\name{calculateAllCutoff}
\alias{calculateAllCutoff}
\title{The function calculate cut off values for all core genes in the set}
\usage{
calculateAllCutoff(coreSet)
}
\arguments{
\item{coreSet}{the path to the core set}
}
\value{
none
}
\description{
The function calculate cut off values for all core genes in the set
}
