% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCoreSet.R
\name{calculateCutoff}
\alias{calculateCutoff}
\title{The function calculate all values, that are necessary for the assessment
process for a specific core gene in the set. The function will saved the
values as the file text in the folder of the core gene}
\usage{
calculateCutoff(root, coreSet, coreGene)
}
\arguments{
\item{root}{the path to the root folder}

\item{coreSet}{the core set name}

\item{coreGene}{the ID of the core gene in the set}
}
\value{
none
}
\description{
The function calculate all values, that are necessary for the assessment
process for a specific core gene in the set. The function will saved the
values as the file text in the folder of the core gene
}
