% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCompleteness.R
\name{checkPreProcess}
\alias{checkPreProcess}
\title{Check if the core set was processed}
\usage{
checkPreProcess(coreSet)
}
\arguments{
\item{coreSet}{the path to the core set}
}
\value{
TRUE or FALSE
}
\description{
Check if the core set was processed
}
