% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCompleteness.R
\name{checkPreProcess}
\alias{checkPreProcess}
\title{Check if the core set was processed with the function processCoreSet()}
\usage{
checkPreProcess(root, coreSet)
}
\arguments{
\item{root}{The path to the core directory, where the core set is stored
within weight_dir, blast_dir, etc.}

\item{coreSet}{The name of the interested core set. The core directory can
contains more than one core set and the user must specify the interested
core set. The core set will be stored in the folder core_orthologs in
subfolder, specify them by the name of the subfolder}
}
\value{
TRUE or FALSE
}
\description{
Check if the core set was processed with the function processCoreSet()
}
\examples{
coreFolder <- system.file("extdata", "sample", package = "fCAT")
checkPreProcess(coreFolder, "test")
}
