% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHamstr.R
\name{concanateFiles}
\alias{concanateFiles}
\title{After HaMStR searched ortholog for the interested genome, it will store the
output in a temporary folder in the core set. The function will concanate all
the output files in the folder in a single files}
\usage{
concanateFiles(directory, genomeName)
}
\arguments{
\item{directory}{path to the folder, that contains the output files}

\item{genomeName}{the genomeID of the genome, that need to be extracted}
}
\value{
none
}
\description{
After HaMStR searched ortholog for the interested genome, it will store the
output in a temporary folder in the core set. The function will concanate all
the output files in the folder in a single files
}
