% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFiles.R
\name{correctDomains}
\alias{correctDomains}
\title{remove lines of a specific genome from a domains file}
\usage{
correctDomains(domainsFile, genome)
}
\arguments{
\item{domainsFile}{path to the file}

\item{genome}{name of the genome in the file, that should be remove}
}
\value{
none
}
\description{
remove lines of a specific genome from a domains file
}
