% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFiles.R
\name{correctFasta}
\alias{correctFasta}
\title{remove all the lines of a specific genome from a fasta file}
\usage{
correctFasta(fasta, genome)
}
\arguments{
\item{fasta}{the path to the fasta file}

\item{genome}{the name of the genome, that should be removed}
}
\value{
none
}
\description{
remove all the lines of a specific genome from a fasta file
}
