% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFiles.R
\name{correctFiles}
\alias{correctFiles}
\title{This function run correctPP, correctFasta, correctDomains, correctPriority 
and correctReport in a loop to correct all phyloprofile, extended fasta, 
domains, prioritylist and report file in a folder}
\usage{
correctFiles(directory, genome)
}
\arguments{
\item{directory}{the path to the directory, that contains the files}

\item{genome}{name of the genome, that should be removed}
}
\value{
none
}
\description{
This function run correctPP, correctFasta, correctDomains, correctPriority 
and correctReport in a loop to correct all phyloprofile, extended fasta, 
domains, prioritylist and report file in a folder
}
