% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFiles.R
\name{correctPriority}
\alias{correctPriority}
\title{If user choose the option to extend the phylogenetic profile of the
interested genome into the original phylogenetic profile, the tool must save
the priority list in a file text in the core directory, so that the tool can
recall it to assess the completeness of the genome, which is appended in the 
original phylogenetic profile. When the user want to redo the check, it can 
be that the user can change the priority list, because of that, anytime a
redo is on, the tool must first remove the old priority list of the genome in
the file.}
\usage{
correctPriority(priorityFile, genome)
}
\arguments{
\item{priorityFile}{The path to the prioritylist file, which contains the
priority lists, which were used for the completeness checking of the genomes,
that were checked with the tool over the time.}

\item{genome}{The genome ID of the genome, that need to be removed from the
prioritylist file}
}
\value{
none
}
\description{
If user choose the option to extend the phylogenetic profile of the
interested genome into the original phylogenetic profile, the tool must save
the priority list in a file text in the core directory, so that the tool can
recall it to assess the completeness of the genome, which is appended in the 
original phylogenetic profile. When the user want to redo the check, it can 
be that the user can change the priority list, because of that, anytime a
redo is on, the tool must first remove the old priority list of the genome in
the file.
}
\examples{
## Create a pseudo priority file with 2 genomes
genomeID <- c("HUMAN@9606@3", "AMPQU@400682@2")
priority_list <- c("HUMAN@9606@3", "AMPQU@400682@2")
table <- data.frame(genomeID, priority_list)

## write it in a file in the current working directory
wd <- getwd()
filePath <- paste(wd, "/fCAT_functiontest.prioritylist", sep = "")
write.table(table, filePath, sep = "\t", row.names = FALSE, quote = FALSE)

## Correcting
correctPriority(filePath, "HUMAN@9606@3")

## Check if HUMAN@9606@3 line is removed from the file
table <- read.table(filePath, sep = "\t", header = TRUE)
print.data.frame(table)

## delete the file
file.remove(filePath)
}
