% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFiles.R
\name{correctReport}
\alias{correctReport}
\title{For every with the tool checked genomes, fCAT will save the frequency table 
of each genome into a file with ending .report in the phyloprofile folder, 
which by default is the output folder in core directory or can be the
folder specified by the user with the argument ppDir in the function 
checkCompleteness. When redo is on to recheck for a genome, whose pp exists
already in the original pp, the tool must remove the lines of the
genome from the original frequency table. That is the purpose of this 
function}
\usage{
correctReport(reportFile, genome)
}
\arguments{
\item{reportFile}{The path to the reprot file (or the frequency table), which
contains the priority lists, which were used for the completeness checking of
the genomes, that were checked with the tool over the time.}

\item{genome}{the genome ID that need to be removed}
}
\value{
none
}
\description{
For every with the tool checked genomes, fCAT will save the frequency table 
of each genome into a file with ending .report in the phyloprofile folder, 
which by default is the output folder in core directory or can be the
folder specified by the user with the argument ppDir in the function 
checkCompleteness. When redo is on to recheck for a genome, whose pp exists
already in the original pp, the tool must remove the lines of the
genome from the original frequency table. That is the purpose of this 
function
}
\examples{
## Create pseudo frequency table file
genomeID <- c("HUMAN@9606@3", "AMPQU@400682")
similar <- c(330, 313)
dissimilar <- c(3, 0)
missing <- c(4, 11)
duplicated <- c(1, 0)
ignored <- c(8, 22)

table <- data.frame(genomeID, similar, dissimilar, missing, duplicated, 
ignored)

## Write the table in a file in the current working directory
wd <- getwd()
filePath <- paste(wd, "/fCAT_functiontest.report", sep = "")
write.table(table, filePath, sep = "\t", row.names = FALSE, quote = FALSE)

correctReport(filePath, "HUMAN@9606@3")
## Check if the file  is corrected
read.table(filePath, sep = "\t", header = TRUE)

## Delete the file
file.remove(filePath)
}
