% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCoreSet.R
\name{createAllAnnotation}
\alias{createAllAnnotation}
\title{The function creates all FAS annotation for all core genes in the set}
\usage{
createAllAnnotation(root, coreSet)
}
\arguments{
\item{root}{the path to the root folder}

\item{coreSet}{the core set name}
}
\value{
none
}
\description{
The function creates all FAS annotation for all core genes in the set
}
