% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenome.R
\name{extractDomains}
\alias{extractDomains}
\title{extract from a domains table a specific genome}
\usage{
extractDomains(domains, genome)
}
\arguments{
\item{domains}{the domains table}

\item{genome}{name of the genome in the file, that should be remove}
}
\value{
the new domains table
}
\description{
extract from a domains table a specific genome
}
