% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenome.R
\name{extractFasta}
\alias{extractFasta}
\title{extract from an extended.fasta file a specific genome}
\usage{
extractFasta(lines, genome)
}
\arguments{
\item{lines}{a vector that contains lines of an extended.fasta file}

\item{genome}{the name of the genome, that should be removed}
}
\value{
the new vector that contains the lines of the genome
}
\description{
extract from an extended.fasta file a specific genome
}
