% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenome.R
\name{extractFasta}
\alias{extractFasta}
\title{This function extract the sequences of a specific genome from a mulitple 
fasta file}
\usage{
extractFasta(lines, genome)
}
\arguments{
\item{lines}{a vector that contains lines of an extended.fasta file}

\item{genome}{the name of the genome, that should be extracted}
}
\value{
the vector, which contains the sequences of the interested genome
}
\description{
This function extract the sequences of a specific genome from a mulitple 
fasta file
}
\examples{
## Create pseudo extended fasta file
fasta <- c(">1001705at2759|HUMAN@9606@3|Q15291|1",
"MNLELLESFGQNYPEEADGTLDCISMALTCTFNRWGT", 
">1489230at2759|ARATH@3702@2|3702_0:005bc2|1",
"MAGRATIPARNSALIAMIADEDTVVGFLMAGVGNVDIRRKTNYLIVDS")
## Extract extended fasta of HUMAN@96063
newFasta <- extractFasta(fasta, "HUMAN@9606@3")
print(newFasta)
}
