% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGenome.R
\name{extractPP}
\alias{extractPP}
\title{This function takes a phylogenetic profile, which contains many different
genomes as an input and extract the phylogenetic profile of a specific genome}
\usage{
extractPP(pp, genome)
}
\arguments{
\item{pp}{A phylogenetic profile in data.frame}

\item{genome}{The genome ID of the genome, whose pp need to be extracted.
Exp: genome = HUMAN@9606@3}
}
\value{
The phylogenetic profile of the interested genome in data.frame
}
\description{
This function takes a phylogenetic profile, which contains many different
genomes as an input and extract the phylogenetic profile of a specific genome
}
