% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{extractPP}
\alias{extractPP}
\title{The function extract a phylogenetic profile of a genome from a phylogentic
profile}
\usage{
extractPP(pp, genome)
}
\arguments{
\item{pp}{the data frame of the phylogenetic profile}

\item{genome}{the ID of the extracted genome}
}
\value{
the phylogenetic profile of the interested genome
}
\description{
The function extract a phylogenetic profile of a genome from a phylogentic
profile
}
