% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{filterIgnore}
\alias{filterIgnore}
\title{Determine if a core gene was ignored by the tool because of the unknown
references species}
\usage{
filterIgnore(coreSet, coreGene, priorityList)
}
\arguments{
\item{coreSet}{the path to the core set}

\item{coreGene}{the ID of the core gene}

\item{priorityList}{the priority list to determine the references species}
}
\value{
TRUE or FALSE
}
\description{
Determine if a core gene was ignored by the tool because of the unknown
references species
}
