% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{filterIgnore}
\alias{filterIgnore}
\title{Determine if a core gene was ignored by the tool because of the unknown
references species}
\usage{
filterIgnore(root, coreSet, coreGene, priorityList)
}
\arguments{
\item{root}{The path to the core directory, where the core set is stored 
within weight_dir, blast_dir, etc.}

\item{coreSet}{The name of the interested core set. The core directory can 
contains more than one core set and the user must specify the interested 
core set. The core set will be stored in the folder core_orthologs in 
subfolder, specify them by the name of the subfolder}

\item{coreGene}{the ID of the core gene}

\item{priorityList}{A list contains one or many genome ID of the genomes,
which were used to build the core set. The genome ID of this list will be 
stored with an priority order, the tool look at into the fasta file of each
core group and determine with the priority order to determine the references
species for each core group.}
}
\value{
TRUE or FALSE
}
\description{
Determine if a core gene was ignored by the tool because of the unknown
references species
}
\examples{
coreFolder <- system.file("extdata", "sample", package = "fCAT")
check <- filterIgnore(coreFolder, "test", "530670", c("HUMAN@9606@1"))
print(check)
}
