% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCompleteness.R
\name{handleError}
\alias{handleError}
\title{Remove the data of a failed check}
\usage{
handleError(coreSet)
}
\arguments{
\item{coreSet}{the path to the core set}
}
\value{
none
}
\description{
Remove the data of a failed check
}
