% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeSeed.R
\name{placeSeed}
\alias{placeSeed}
\title{The function move the fasta file and the annotation file of the interested
genome in to the equivalent folder in the core set to perfome the assessment}
\usage{
placeSeed(genome, fasAnno = NULL, root, process = FALSE)
}
\arguments{
\item{genome}{the path to the fasta file of the genome}

\item{fasAnno}{the path to the fas annotation file of the genome. If equal
NULL, the function will compute the annotation}

\item{root}{the path to the root folder}

\item{process}{A logical option to determine if the function was used as a
subfunction in the processCoreSet function}
}
\value{
none
}
\description{
The function move the fasta file and the annotation file of the interested
genome in to the equivalent folder in the core set to perfome the assessment
}
