% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeReport.R
\name{printPriority}
\alias{printPriority}
\title{Add the priority list of the interested genome to the priority file. If the
file doesnt exist it will be created}
\usage{
printPriority(genomeID, priorityList, coreSet, scoreMode)
}
\arguments{
\item{genomeID}{the genomeID of the interested genome}

\item{priorityList}{the priority list}

\item{coreSet}{the path to the core set}

\item{scoreMode}{the mode determines the method to assess the founded
ortholog}
}
\value{
none
}
\description{
Add the priority list of the interested genome to the priority file. If the
file doesnt exist it will be created
}
