% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeReport.R
\name{printReport}
\alias{printReport}
\title{Create the report of the completeness of the interested genome with the other
genome in the original phylogenetic profile}
\usage{
printReport(report, coreSet, scoreMode)
}
\arguments{
\item{report}{the data frame that contains the report}

\item{coreSet}{the path to the coreSet}

\item{scoreMode}{the mode determines the way to assess the founded ortholog}
}
\value{
none
}
\description{
Create the report of the completeness of the interested genome with the other
genome in the original phylogenetic profile
}
