% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCoreSet.R
\name{processCoreSet}
\alias{processCoreSet}
\title{The function compute the FAS annotations and calculate the cut off values for
all core genes in the set}
\usage{
processCoreSet(coreSet)
}
\arguments{
\item{coreSet}{the path to the core set}
}
\value{
none
}
\description{
The function compute the FAS annotations and calculate the cut off values for
all core genes in the set
}
