% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCoreSet.R
\name{processCoreSet}
\alias{processCoreSet}
\title{The function run createAllAnnotation and calculateAllCufoff to process
the core set}
\usage{
processCoreSet(coreDir, coreSet)
}
\arguments{
\item{coreDir}{The path to the core directory, where the core set is stored
within weight_dir, blast_dir, etc.}

\item{coreSet}{The name of the interested core set. The core directory can
contains more than one core set and the user must specify the interested
core set. The core set will be stored in the folder core_orthologs in
subfolder, specify them by the name of the subfolder}
}
\value{
none
}
\description{
The function run createAllAnnotation and calculateAllCufoff to process
the core set
}
\examples{
coreFolder <- system.file("extdata", "sample", package = "fCAT")
processCoreSet(coreFolder, "test")
}
