% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHamstr.R
\name{runHamstr}
\alias{runHamstr}
\title{This function takes a path to a core set and run HaMStR to search ortholog on
the genome in the folder genome_dir of core set}
\usage{
runHamstr(coreSet, extend = FALSE, scoreMode, priorityList = NULL, cpu)
}
\arguments{
\item{coreSet}{the path to the core set}

\item{extend}{if extend=TRUE the phylogenetic profile of the genome will be
appended to the original phylogenetic profile}

\item{scoreMode}{the mode determines the way to assess the founded ortholog}

\item{priorityList}{the list determines the references species}

\item{cpu}{determines the cores that HaMStR will use}
}
\value{
phylogenetic profile of the genome
}
\description{
This function takes a path to a core set and run HaMStR to search ortholog on
the genome in the folder genome_dir of core set
}
