% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{translateReport}
\alias{translateReport}
\title{Translate the report table into a frequent table, how many is complete, etc.}
\usage{
translateReport(genomeID, report, scoreMode)
}
\arguments{
\item{genomeID}{the genome ID of the interested genome}

\item{report}{the report in data frame}

\item{scoreMode}{the mode to determine the method to assess the ortholog}
}
\value{
A frequencies table in data frame
}
\description{
Translate the report table into a frequent table, how many is complete, etc.
}
