% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busco.R
\name{updateLength}
\alias{updateLength}
\title{In the scoreMode "busco" the tool must calculate the length of the founded
ortholog to perfome the assessment. After HaMStR founded an ortholog the
function will be called to update the length of the ortholog in phylogenetic
profile}
\usage{
updateLength(root, coreSet, coreGene, genomeName)
}
\arguments{
\item{root}{the path to the root folder}

\item{coreSet}{the core set name}

\item{coreGene}{the ID of the core gene}

\item{genomeName}{the genome ID of the interested genome}
}
\value{
none
}
\description{
In the scoreMode "busco" the tool must calculate the length of the founded
ortholog to perfome the assessment. After HaMStR founded an ortholog the
function will be called to update the length of the ortholog in phylogenetic
profile
}
